Enable polymorphism in overloaded operators
Rather than relying on a static return type pre-defined for each operator,
instead, just type-check the (now-resolved) function.  As long as the functions
are loaded into gamma with the right polymorphic types (NOT DONE YET), this
lets you have polymorphic overloaded operators, which we already supported for
inputs, but not for their outputs.

git-svn-id: https://frenetic-lang.org/svn/frenetic/src/compiler@7137 41643c41-6115-4a67-97d8-b2dbeab842be